Turn off encryption after halting
Turning off encryption MUST happen after the command has been
sent. Because of this the card is never halted. This is what is
causing the same card to be read over and over again.

Please note that if the card is halted, the status returned will be ERR
- the card it halted and so can not respond. so a timeout will occur. In
this instance, this is the correct response.

Fixes #9.